patch 9.0.2002: Vim9: need cleanup of class related interface code
Problem:  Vim9: need cleanup of class related interface code
Solution: Remove the unused class variable and class method related code
          for interfaces.

Remove unused class variable and class method related code for
interfaces.

Refactor the code.

Optimize the object/class member double lookup in compile_lhs().

Change unused global functions to static functions.

closes: #13302

+++ src/vim9cmds.c
@@ -254,7 +254,7 @@ compile_lock_unlock(
 		{
 		    // Push the class of the bare class variable name
 		    name = cl->class_name;
-		    len = STRLEN(name);
+		    len = (int)STRLEN(name);
 #ifdef LOG_LOCKVAR
 		    ch_log(NULL, "LKVAR:    ... cctx_class_member: name %s",
 			   name);
